refactor: ThemeDetail component
